Doing instruction parsing differently + vm_decode

+++ include/vcvm/bytecode.h
@@ -31,7 +31,7 @@ typedef struct bytecode
     intrinsic_t** intrinsics;
 
     uint32_t instruction_count;
-    uint32_t* instructions;
+    uint8_t* instructions;
 } bytecode_t;
 
 int bytecode_init(bytecode_t* bytecode, allocator_options_t* allocator_options, const uint8_t* raw, size_t size, resolve_intrinsic_delegate_t resolve);
